+++ app/javascript/packs/shared/navigation/components/NotificationImage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const NotificationImage = ({type_of}) => {
+  return(
+    <div className="text-center">
+      <span className="assignment">
+        <i className="fa fa-newspaper-o" />
+      </span>
+    </div>
+  )
+}
+
+export default NotificationImage;
