+++ app/javascript/packs/shared/reducers/UsersReducer.js
@@ -1,4 +1,4 @@
-import { SET_CURRENT_USER } from "../actions/types";
+import { SET_CURRENT_USER } from "../../app/action_types";
 
 export function currentUser(
   state = { id: 0, fullName: "", avatarPath: "" },
